+++ fe1-web/features/events/components/Event.tsx
@@ -9,8 +9,8 @@ import ParagraphBlock from 'components/ParagraphBlock';
 import TextBlock from 'components/TextBlock';
 import { Hash, Timestamp } from 'model/objects';
 import { EventMeeting } from 'features/meeting/components';
-import EventElection from 'features/evoting/components/EventElection';
-import EventRollCall from 'features/rollCall/components/EventRollCall';
+import { EventElection } from 'features/evoting/components';
+import { EventRollCall } from 'features/rollCall/components';
 import { Meeting } from 'features/meeting/objects';
 import { Election } from 'features/evoting/objects';
 import { RollCall } from 'features/rollCall/objects';
