+++ fe1-web/features/events/reducer/index.ts
@@ -1,0 +1,2 @@
+export { default as eventsReducer } from './EventsReducer';
+export * from './EventsReducer';
