+++ fe1-web/features/events/screens/index.ts
@@ -1,0 +1,1 @@
+export { default as CreateEvent } from './CreateEvent';
