+++ fe1-web/features/evoting/components/index.ts
@@ -1,0 +1,1 @@
+export { default as EventElection } from './EventElection';
