+++ fe1-web/features/evoting/screens/index.ts
@@ -1,0 +1,1 @@
+export { default as CreateElection } from './CreateElection';
