--- fe1-web/features/lao/screens/Organizer.tsx
@@ -1,14 +1,0 @@
-import React from 'react';
-
-import Attendee from './Attendee';
-
-/**
- * Organizer screen: lists editable LAO properties and past/ongoing/future events
- *
- * Currently just a simple clone of the Attendee screen
- */
-const Organizer = () => (
-  <Attendee />
-);
-
-export default Organizer;
