+++ fe1-web/features/lao/screens/OrganizerScreen.tsx
@@ -1,0 +1,14 @@
+import React from 'react';
+
+import AttendeeScreen from './AttendeeScreen';
+
+/**
+ * OrganizerScreen: lists editable LAO properties and past/ongoing/future events
+ *
+ * Currently just a simple clone of the AttendeeScreen screen
+ */
+const OrganizerScreen = () => (
+  <AttendeeScreen />
+);
+
+export default OrganizerScreen;
