+++ fe1-web/features/lao/screens/index.ts
@@ -1,0 +1,3 @@
+export { default as Attendee } from 'features/lao/screens/AttendeeScreen';
+export { default as Organizer } from 'features/lao/screens/OrganizerScreen';
+export { default as Witness } from 'features/lao/screens/WitnessScreen';
