+++ fe1-web/features/meeting/network/MeetingHandler.ts
@@ -2,7 +2,7 @@ import { ExtendedMessage } from 'model/network/method/message';
 import { ActionType, MessageRegistry, ObjectType } from 'model/network/method/message/data';
 import { dispatch, getStore, makeCurrentLao } from 'store';
 import { hasWitnessSignatureQuorum } from 'ingestion/handlers/Utils';
-import { addEvent, updateEvent } from 'features/events/reducer/EventsReducer';
+import { addEvent, updateEvent } from 'features/events/reducer';
 import { getEventFromId } from 'features/events/network/EventHandlerUtils';
 
 import { CreateMeeting, StateMeeting } from './messages';
