+++ fe1-web/features/rollCall/components/index.ts
@@ -1,0 +1,1 @@
+export { default as EventRollCall } from './EventRollCall';
