+++ fe1-web/features/rollCall/network/RollCallHandler.ts
@@ -9,7 +9,7 @@ import {
   setLaoLastRollCall,
 } from 'store';
 import { subscribeToChannel } from 'network/CommunicationApi';
-import { addEvent, updateEvent } from 'features/events/reducer/EventsReducer';
+import { addEvent, updateEvent } from 'features/events/reducer';
 import { getEventFromId } from 'features/events/network/EventHandlerUtils';
 import * as Wallet from 'features/wallet/objects';
