+++ fe1-web/features/rollCall/screens/index.ts
@@ -1,0 +1,2 @@
+export { default as CreateRollCall } from './CreateRollCall';
+export { default as RollCallOpened } from './RollCallOpened';
