--- fe1-web/features/social/navigation/index.ts
@@ -1,2 +1,0 @@
-export { default as SocialMediaNavigation } from './SocialMediaNavigation';
-export { default as SocialSearchNavigation } from './SocialSearchNavigation';
