+++ fe1-web/features/social/screens/SocialSearch.tsx
@@ -10,7 +10,7 @@ import { makeCurrentLao } from 'store';
 import { PublicKey } from 'model/objects';
 import TextBlock from 'components/TextBlock';
 import { gray } from 'styles/colors';
-import { makeRollCallAttendeesList } from 'features/events/reducer/EventsReducer';
+import { makeRollCallAttendeesList } from 'features/events/reducer';
 
 import { UserListItem } from '../components';
