--- fe1-web/features/wallet/navigation/index.ts
@@ -1,1 +1,0 @@
-export { default as WalletNavigation } from './WalletNavigation';
