+++ fe1-web/navigation/bars/LaoNavigation.tsx
@@ -11,9 +11,9 @@ import { PublicKey } from 'model/objects';
 import STRINGS from 'res/strings';
 import Home from 'parts/Home';
 import Identity from 'parts/lao/Identity';
-import Attendee from 'features/lao/screens/Attendee';
-import { SocialMediaNavigation } from 'features/social/navigation';
-import { WalletNavigation } from 'features/wallet/navigation';
+import { Attendee } from 'features/lao/screens';
+import SocialMediaNavigation from 'features/social/navigation/SocialMediaNavigation';
+import WalletNavigation from 'features/wallet/navigation/WalletNavigation';
 import OrganizerNavigation from './organizer/OrganizerNavigation';
 import WitnessNavigation from './witness/WitnessNavigation';
