+++ fe1-web/navigation/bars/witness/WitnessNavigation.tsx
@@ -3,12 +3,12 @@ import { createStackNavigator } from '@react-navigation/stack';
 
 import STRINGS from 'res/strings';
 
-import Witness from 'features/lao/screens/Witness';
+import { Witness } from 'features/lao/screens';
 import WitnessCamera from 'parts/lao/witness/WitnessCamera';
 
 /**
- * Define the Witness stack navigation
- * Allows to navigate to the Witness and the WitnessCamera screen
+ * Define the WitnessScreen stack navigation
+ * Allows to navigate to the WitnessScreen and the WitnessCamera screen
 */
 
 const Stack = createStackNavigator();
