+++ fe1-web/res/strings.ts
@@ -27,12 +27,12 @@ const STRINGS = {
   organization_navigation_tab_witness: 'Witness',
   organization_navigation_tab_identity: 'My identity',
 
-  /* --- Witness Navigation Strings --- */
+  /* --- WitnessScreen Navigation Strings --- */
   witness_navigation_tab_home: 'Witness',
-  witness_navigation_tab_video: 'Witness Video',
+  witness_navigation_tab_video: 'WitnessScreen Video',
 
-  /* --- Organizer Navigation Strings --- */
-  organizer_navigation_tab_home: 'Organizer Home',
+  /* --- OrganizerScreen Navigation Strings --- */
+  organizer_navigation_tab_home: 'OrganizerScreen Home',
   organizer_navigation_tab_create_event: 'Create Event',
   organizer_navigation_tab_add_witness: 'Add witness',
   organizer_navigation_tab_roll_call: 'Roll-Call',
@@ -100,7 +100,7 @@ const STRINGS = {
   launch_address: 'Address',
   launch_button_launch: 'Launch',
 
-  /* --- Organizer Strings --- */
+  /* --- OrganizerScreen Strings --- */
   organization_name: 'Organization name',
 
   /* --- Identity Strings --- */
@@ -109,7 +109,7 @@ const STRINGS = {
   identity_check_box_anonymous_description: 'You can participate in organizations and meetings anonymously by leaving '
     + 'this box checked. If you wish to reveal your identity to other participants in the organization, you may '
     + 'un-check this box and enter the information you wish to reaveal below. You must enter identity information in '
-    + 'order to play an Organizer or Witness role in an organization.',
+    + 'order to play an OrganizerScreen or WitnessScreen role in an organization.',
   identity_name_placeholder: 'Name',
   identity_title_placeholder: 'Title',
   identity_organization_placeholder: 'Organization',
@@ -117,8 +117,8 @@ const STRINGS = {
   identity_phone_placeholder: 'Phone number',
   identity_qrcode_description: 'ID (Public Key):',
 
-  /* --- Witness Strings --- */
-  witness_description: 'Witness screen',
+  /* --- WitnessScreen Strings --- */
+  witness_description: 'WitnessScreen screen',
   witness_video_button: 'Go to the video screen',
   witness_name: 'Witnesses',
   witness_scan: 'Please scan the personal QR code of the witness to add',
@@ -173,7 +173,7 @@ const STRINGS = {
   roll_call_modal_enter_token: 'Enter token:',
   roll_call_participant_added: 'participant added',
   roll_call_invalid_token: 'invalid participant token',
-  roll_call_attendee_token_placeholder: 'Attendee token',
+  roll_call_attendee_token_placeholder: 'AttendeeScreen token',
 
   /* --- Poll creation Strings --- */
   poll_create_question: 'Question*',
