+++ fe1-web/store/reducers/RootReducer.ts
@@ -2,7 +2,7 @@ import AsyncStorage from '@react-native-async-storage/async-storage';
 import { persistCombineReducers } from 'redux-persist';
 
 import { socialReducer } from 'features/social/reducer';
-import eventsReducer from 'features/events/reducer/EventsReducer';
+import { eventsReducer } from 'features/events/reducer';
 import { walletReducer } from 'features/wallet/reducer';
 import laoReducer from './LaoReducer';
 import keyPairReducer from './KeyPairReducer';
